fix find_command_in_pkg on windows
